requirements.txt
numpy
opencv-python-headless
torch>=2.3.0
torchvision
torchaudio
diffusers>=0.27.0
transformers>=4.40.0
safetensors
accelerate>=0.30.0
huggingface_hub
pillow
streamlit==1.32.0
scipy
python-dotenv
controlnet-aux
basicsr
facexlib
xformers

# Note: Photorealistic Refinement (Phase 2) is automatically initialized via SDXL ControlNet
# and runs on native PyTorch/Diffusers. All weights are fetched to NVMe automatically.